game+gui: audit tail -- stall publish, the invented video fallback, the  VideoDefinition layout, two tripwires
Five reviewed findings, none of them large on its own, one of which turned out
to be hiding a live bug.

F-P3-3 (S2) -- THE STALL PUBLISH. ResourceUpdateThread @0x823BC9B8 ends:
    ret = (gameDataModule->vtable+0x44)(...)
    cmpwi cr6, r3, 1 ; li r10,1 ; beq ; li r10,0
    stbx r10, r31, 0x99FFE8            ; (gm+1600)+0x99FFE8 == gm+0x9A0628
That byte is what ConstructUpdateSetFromFsm turns into update-set bit 0x400,
the STALL bit. We threw the return value away and never wrote mbStalled, so
the whole stall gate was dead -- nothing downstream could ever tell that the
streamer had not caught up. Republished every pass, not latched, as on the
console.

F-P8b-3 (S4) -- THE INVENTED VIDEO FALLBACK, and what it was hiding.
MovieManager::AcquireVideoDataResource fell back to
FindFirstResourceOfType(E_RESOURCETYPE_VIDEODATA) when the id lookup missed.
No console analogue exists: the acquire is strictly BY HANDLE
(PendingVideoDataResourceRequest @0x824F7808, which asserts IsMovieQueued ->
CreateFromHandle @0x82507DD4); there is no type scan anywhere in the image.
It did not make a miss survivable, it made it invisible.

Deleting it immediately surfaced a real miss: EAFranchise came back NOT FOUND
and had been playing only because the fallback handed it whatever was first in
the pool. Cause below.

F-P8b-6 (S5) -- THE VideoDefinition LAYOUT, read store-for-store off
VideoDefinition::Prepare @0x82472990:
    std  ID::HashString("")  , 0x10(this)      <- ONE 8-byte id
    stw  [dword_830082A8]    , 0x18(this)      <- the SOUND-NAME default
    stw  0, 0x1C ; stw 0, 0x20                 <- the TWO crossfades
    stb  0, 0x24/0x25/0x26
We split the 64-bit id into {id, soundName} at +0x10/+0x14 and then slid every
later field back a slot: "miCrossfadeInFrames" was the sound name,
"miCrossfadeOutFrames" was the first crossfade, "muField20" was the second.
Corrected, and dword_830082A8's image initialiser (0x0233B6D5) is now a named
mutable global rather than a hardcoded 0. GuiEventPlayVideo::muVideoResourceId
widened to u64 to match, and the queued definition now carries the sound name
across HandlePlayVideoEvent -- filling it in the producer was pointless while
the first copy dropped it.

...which is where EAFranchise's miss came from: our HashString returns s32,
and widening it directly sign-extends 0xF4DE3E5C to 0xFFFFFFFFF4DE3E5C, which
matches nothing. The console's hasher ends `clrldi r3,r11,32` -- the top half
is always zero. Zero-extended at both sites now, and both logos acquire:
    [MovieManager] acquire id=0x00000000F4DE3E5C entry=found idx=4
    [MovieManager] acquire id=0x000000007C822271 entry=found idx=1

F-P8b-14 (S5, ledger drift) -- ScreenLoading::GetResourcesToLoad cited the ICF
fold @0x824B5A20 and reproduced that body: two "Invalid pointer" asserts and a
null tuple store. Wrong slot by one -- 0x824B5A20 is the end of the PREVIOUS
vtable. Slot +0x20 of vtable @0x82074130 is 0x825011B0, three instructions:
    li r11, 0 ; stw r11, 0(r5) ; blr
It writes the COUNT only and leaves the caller's tuple pointer alone. Both
asserts and the tuple store were invented.

F-P8b-12 (S5) -- BootPreload::OnLeave's missing mpGuiCache assert. One line,
but it is the one place that would say the preload ladder above ran against
nothing.

Boot verified green throughout: both logos acquire and play with audio, no
asserts, no exceptions.

+++ src/GameSource/Gui/BrnGuiVideoEvents.h
@@ -24,7 +24,10 @@ namespace BrnGui
     // queued movie (DecFIGS BrnGui::GuiEventPlayVideo).
     struct GuiEventPlayVideo : public CgsGui::GuiEvent<KI_GUIEVENT_PLAY_VIDEO>
     {
-        u32  muVideoResourceId;            // VideoDataResource id to load + play
+        // ⭐ WIDENED 2026-08-16 (boot audit F-P8b-5/F-P8b-6): the producers store this with
+        // `std` at VideoDefinition+0x10 -- it is one 64-bit CgsResource::ID, and the PC was
+        // truncating it to u32 on the way in and back out again.
+        u64  muVideoResourceId;            // +0x10 VideoDataResource id to load + play
         f32  mafRectangle[4];              // left, top, right, bottom (logical 1280x720) -- X360 Vector4
         s32  miCrossfadeInFrames;
         s32  miCrossfadeOutFrames;
